+++ src/main/java/com/imsweb/layout/Field.java
@@ -3,6 +3,8 @@
  */
 package com.imsweb.layout;
 
+import java.util.Objects;
+
 /**
  * A single field from the layout.
  * <p/>
@@ -58,49 +60,34 @@ public class Field {
         }
     }
 
-    /**
-     * Field name (required)
-     */
+    // field name (required)
     protected String _name;
 
-    /**
-     * Short label (required)
-     */
+    // short label (required)
     protected String _shortLabel;
 
-    /**
-     * Long label (required)
-     */
+    // long label (required)
     protected String _longLabel;
 
-    /**
-     * NAACCR Item Number (optional)
-     */
+    // field length (optional, depends on the type of layout)
+    protected Integer _length;
+
+    // NAACCR Item Number (optional)
     protected Integer _naaccrItemNum;
 
-    /**
-     * Alignment (default to left)
-     */
+    // alignment (default to left)
     protected FieldAlignment _align;
 
-    /**
-     * Padding character (optional)
-     */
+    // padding character (optional)
     protected String _padChar;
 
-    /**
-     * Default Value (default to null)
-     */
+    // default Value (default to null)
     protected String _defaultValue;
 
-    /**
-     * Trimming (defaults to true)
-     */
+    // trimming (defaults to true)
     protected Boolean _trim;
 
-    /**
-     * Section of field (optional)
-     */
+    // section of field (optional)
     protected String _section;
 
     /**
@@ -110,128 +97,82 @@ public class Field {
         _trim = Boolean.TRUE;
     }
 
-    /**
-     * @return Returns the name.
-     */
     public String getName() {
         return _name;
     }
 
-    /**
-     * @param name The name to set.
-     */
     public void setName(String name) {
         this._name = name;
     }
 
-    /**
-     * @return Returns the shortLabel.
-     */
     public String getShortLabel() {
         return _shortLabel;
     }
 
-    /**
-     * @param shortLabel The shortLabel to set.
-     */
     public void setShortLabel(String shortLabel) {
         this._shortLabel = shortLabel;
     }
 
-    /**
-     * @return Returns the longLabel.
-     */
     public String getLongLabel() {
         return _longLabel;
     }
 
-    /**
-     * @param longLabel The longLabel to set.
-     */
     public void setLongLabel(String longLabel) {
         this._longLabel = longLabel;
     }
 
-    /**
-     * @return Returns the naaccrItemNum.
-     */
+    public Integer getLength() {
+        return _length;
+    }
+
+    public void setLength(Integer length) {
+        _length = length;
+    }
+
     public Integer getNaaccrItemNum() {
         return _naaccrItemNum;
     }
 
-    /**
-     * @param naaccrItemNum The naaccrItemNum to set.
-     */
     public void setNaaccrItemNum(Integer naaccrItemNum) {
         this._naaccrItemNum = naaccrItemNum;
     }
 
-    /**
-     * @return Returns the align.
-     */
     public FieldAlignment getAlign() {
         return _align;
     }
 
-    /**
-     * @param align The align to set.
-     */
     public void setAlign(FieldAlignment align) {
         this._align = align;
     }
 
-    /**
-     * @return Returns the default value.
-     */
     public String getDefaultValue() {
         return _defaultValue;
     }
 
-    /**
-     * @param defaultValue The default value to set.
-     */
     public void setDefaultValue(String defaultValue) {
         this._defaultValue = defaultValue;
     }
 
-    /**
-     * @return Returns the padChar.
-     */
     public String getPadChar() {
         return _padChar;
     }
 
-    /**
-     * @param padChar The padChar to set.
-     */
     public void setPadChar(String padChar) {
         this._padChar = padChar;
     }
 
-    /**
-     * @return Returns the trim value.
-     */
     public Boolean getTrim() {
         return _trim;
     }
 
-    /**
-     * @param trim The trim value to set.
-     */
     public void setTrim(Boolean trim) {
         this._trim = trim;
     }
 
-    /**
-     * @return Returns the section
-     */
     public String getSection() {
         return _section;
     }
 
-    /**
-     * @param section The section value to set
-     */
     public void setSection(String section) {
         _section = section;
     }
@@ -242,28 +183,15 @@ public class Field {
     }
 
     @Override
-    public int hashCode() {
-        int prime = 31;
-        int result = 1;
-        result = prime * result + ((_name == null) ? 0 : _name.hashCode());
-        return result;
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Field that = (Field)o;
+        return Objects.equals(_name, that._name);
     }
 
     @Override
-    public boolean equals(Object obj) {
-        if (this == obj)
-            return true;
-        if (obj == null)
-            return false;
-        if (getClass() != obj.getClass())
-            return false;
-        Field other = (Field)obj;
-        if (_name == null) {
-            if (other._name != null)
-                return false;
-        }
-        else if (!_name.equals(other._name))
-            return false;
-        return true;
+    public int hashCode() {
+        return Objects.hash(_name);
     }
 }